Pre-update cleanup.
