+++ client/scripts/controllers/chain/cosmos/account.ts
@@ -3,8 +3,7 @@ import BN from 'bn.js';
 import { IApp } from 'state';
 import CosmosChain from 'controllers/chain/cosmos/chain';
 import { CosmosToken } from 'controllers/chain/cosmos/types';
-import { Account, IAccountsModule, ITXModalData } from 'models';
-import { AccountsStore } from 'stores';
+import { Account, ITXModalData } from 'models';
 import { AuthAccountsResponse } from '@cosmjs/launchpad';
 import {
   MsgSendEncodeObject,
@@ -13,6 +12,7 @@ import {
   MsgWithdrawDelegatorRewardEncodeObject,
 } from '@cosmjs/stargate';
 import { BondStatus } from '@cosmjs/launchpad/build/lcdapi/staking';
+import CosmosAccounts from './accounts';
 
 export interface ICosmosValidator {
   // TODO: add more properties (commission, unbonding, jailed, etc)
@@ -25,7 +25,7 @@ export interface ICosmosValidator {
   isJailed: boolean;
 }
 
-export class CosmosAccount extends Account<CosmosToken> {
+export default class CosmosAccount extends Account<CosmosToken> {
   private _Chain: CosmosChain;
   private _Accounts: CosmosAccounts;
 
@@ -141,56 +141,5 @@ export class CosmosAccount extends Account<CosmosToken> {
       }
     };
     await this._Chain.sendTx(this, msg);
-  }
-}
-
-export class CosmosAccounts implements IAccountsModule<CosmosToken, CosmosAccount> {
-  private _initialized: boolean = false;
-  public get initialized() { return this._initialized; }
-
-  // STORAGE
-  private _store: AccountsStore<CosmosAccount> = new AccountsStore();
-  public get store() { return this._store; }
-
-  private _Chain: CosmosChain;
-
-  public get(address: string) {
-    return this.fromAddress(address);
-  }
-
-  private _app: IApp;
-  public get app() { return this._app; }
-
-  constructor(app: IApp) {
-    this._app = app;
-  }
-
-  public fromAddress(address: string): CosmosAccount {
-    // accepts bech32 encoded cosmosxxxxx addresses and not cosmospubxxx
-    let acct;
-    try {
-      acct = this._store.getByAddress(address);
-    } catch (e) {
-      acct = new CosmosAccount(this.app, this._Chain, this, address);
-    }
-    return acct;
-  }
-
-  public fromAddressIfExists(address: string): CosmosAccount | null {
-    try {
-      return this._store.getByAddress(address);
-    } catch (e) {
-      return null;
-    }
-  }
-
-  public deinit() {
-    this._initialized = false;
-    this.store.clear();
-  }
-
-  public async init(ChainInfo: CosmosChain): Promise<void> {
-    this._Chain = ChainInfo;
-    this._initialized = true;
   }
 }
