+++ client/scripts/controllers/chain/cosmos/accounts.ts
@@ -1,0 +1,58 @@
+import _ from 'lodash';
+import { IApp } from 'state';
+import CosmosChain from 'controllers/chain/cosmos/chain';
+import { CosmosToken } from 'controllers/chain/cosmos/types';
+import { IAccountsModule } from 'models';
+import { AccountsStore } from 'stores';
+import CosmosAccount from './account';
+
+export default class CosmosAccounts implements IAccountsModule<CosmosToken, CosmosAccount> {
+  private _initialized: boolean = false;
+  public get initialized() { return this._initialized; }
+
+  // STORAGE
+  private _store: AccountsStore<CosmosAccount> = new AccountsStore();
+  public get store() { return this._store; }
+
+  private _Chain: CosmosChain;
+
+  public get(address: string) {
+    return this.fromAddress(address);
+  }
+
+  private _app: IApp;
+  public get app() { return this._app; }
+
+  constructor(app: IApp) {
+    this._app = app;
+  }
+
+  public fromAddress(address: string): CosmosAccount {
+    // accepts bech32 encoded cosmosxxxxx addresses and not cosmospubxxx
+    let acct;
+    try {
+      acct = this._store.getByAddress(address);
+    } catch (e) {
+      acct = new CosmosAccount(this.app, this._Chain, this, address);
+    }
+    return acct;
+  }
+
+  public fromAddressIfExists(address: string): CosmosAccount | null {
+    try {
+      return this._store.getByAddress(address);
+    } catch (e) {
+      return null;
+    }
+  }
+
+  public deinit() {
+    this._initialized = false;
+    this.store.clear();
+  }
+
+  public async init(ChainInfo: CosmosChain): Promise<void> {
+    this._Chain = ChainInfo;
+    this._initialized = true;
+  }
+}
